Adding a Random Text Paragraphs Generator
This is an interactive program (meaning, it does require user inputs like number of paragraphs) to generate Random Text File with Python 3, in the same directory where this script will run. The code is well-commented, and includes exception handling to handle errors.

Note:
The code is contributed as a part of CH20 (Contributor's Hack 2020) by [HakinCodes](hakincodes.tech).