Created a basic user friendly Calculator
The user can type any statement for example if I want to add I can give input like : "Please add the numbers for me"
and the output after adding the particular numbers will be printed.
It is built to make interaction easier and not make it monotous for the user by typing only numerical values as choice rather giving any usual english statements.

It is a better version of Simple Calculator.